add test cases; test_with_remote and test_with_local

+++ test/config/deploy.rb
@@ -23,6 +23,8 @@ set(:java_setup_locally, true)
 
 ## mvn ##
 set(:mvn_tools_path_local, File.expand_path("tmp/mvn"))
+set(:mvn_project_path) { release_path }
+set(:mvn_project_path_local, repository)
 
 role :web, "192.168.33.10"
 role :app, "192.168.33.10"
@@ -84,20 +86,28 @@ def reset_mvn!
   variables.each_key do |key|
     reset!(key) if /^mvn_/ =~ key
   end
+# find_and_execute_task("mvn:setup_default_environment")
 end
 
 def uninstall_mvn!
   run("rm -rf #{mvn_path.dump}")
   run("rm -f #{mvn_archive_file.dump}")
   run_locally("rm -rf #{mvn_path_local.dump}")
-  unless mvn_settings.empty?
-    run("rm -f #{mvn_settings.map { |x| File.join(mvn_settings_path, x).dump }.join(" ")}")
-    run_locally("rm -f #{mvn_settings.map { |x| File.join(mvn_settings_path_local, x).dump }.join(" ")}")
-  end
+  run("rm -f #{mvn_settings.map { |x| File.join(mvn_settings_path, x).dump }.join(" ")}") unless mvn_settings.empty?
+  run_locally("rm -f #{mvn_settings_local.map { |x| File.join(mvn_settings_path_local, x).dump }.join(" ")}") unless mvn_settings_local.empty?
+  run("rm -rf #{mvn_target_path.dump}")
+  run_locally("rm -rf #{mvn_target_path_local.dump}")
+  reset_mvn!
 end
 
 task(:test_all) {
   find_and_execute_task("test_default")
+  find_and_execute_task("test_with_remote")
+  find_and_execute_task("test_with_local")
+}
+
+on(:start) {
+  run("rm -rf #{deploy_to.dump}")
 }
 
 namespace(:test_default) {
@@ -113,20 +123,19 @@ namespace(:test_default) {
     uninstall_mvn!
     set(:mvn_version, "3.0.5")
     set(:mvn_skip_tests, true)
+    set(:mvn_setup_remotely, true)
+    set(:mvn_setup_locally, true)
     set(:mvn_update_remotely, true)
     set(:mvn_update_locally, true)
-#   set(:mvn_project_path) { release_path }
-    set(:mvn_project_path_local, repository)
     set(:mvn_template_path, File.join(File.dirname(__FILE__), "templates"))
     set(:mvn_settings, %w(settings.xml))
+    find_and_execute_task("mvn:setup_default_environment")
     find_and_execute_task("deploy:setup")
     find_and_execute_task("deploy")
   }
 
   task(:teardown) {
-    reset_mvn!
     uninstall_mvn!
-    run("rm -rf #{deploy_to.dump}")
   }
 
   task(:test_run_mvn) {
@@ -136,7 +145,7 @@ namespace(:test_default) {
   }
 
   task(:test_run_mvn_via_sudo) {
-    assert_command("#{mvn_cmd} --version")
+    assert_command("#{mvn_cmd} --version", :via => :sudo)
   }
 
   task(:test_run_mvn_without_path) {
@@ -157,13 +166,141 @@ namespace(:test_default) {
     mvn.exec_locally("--version")
   }
 
-  task(:test_mvn_project) {
-    assert_file_exists(File.join(mvn_project_path, "pom.xml"))
+  task(:test_mvn_artifact) {
     assert_file_exists(File.join(mvn_project_path, "target", "capistrano-maven-0.0.1-SNAPSHOT.jar"))
   }
 
-  task(:test_mvn_project_locally) {
-    assert_file_exists(File.join(mvn_project_path_local, "pom.xml"), :via => :run_locally)
+  task(:test_mvn_artifact_locally) {
+    assert_file_exists(File.join(mvn_project_path_local, "target", "capistrano-maven-0.0.1-SNAPSHOT.jar"), :via => :run_locally)
+  }
+}
+
+namespace(:test_with_remote) {
+  task(:default) {
+    methods.grep(/^test_/).each do |m|
+      send(m)
+    end
+  }
+  before "test_with_remote", "test_with_remote:setup"
+  after "test_with_remote", "test_with_remote:teardown"
+
+  task(:setup) {
+    uninstall_mvn!
+    set(:mvn_version, "3.0.5")
+    set(:mvn_skip_tests, true)
+    set(:mvn_setup_remotely, true)
+    set(:mvn_setup_locally, false)
+    set(:mvn_update_remotely, true)
+    set(:mvn_update_locally, false)
+    set(:mvn_template_path, File.join(File.dirname(__FILE__), "templates"))
+    set(:mvn_settings, %w(settings.xml))
+    find_and_execute_task("mvn:setup_default_environment")
+    find_and_execute_task("deploy:setup")
+    find_and_execute_task("deploy")
+  }
+
+  task(:teardown) {
+    uninstall_mvn!
+  }
+
+  task(:test_run_mvn) {
+    assert_file_exists(mvn_bin)
+    assert_file_exists(File.join(mvn_settings_path, "settings.xml"))
+    assert_command("#{mvn_cmd} --version")
+  }
+
+  task(:test_run_mvn_via_sudo) {
+    assert_command("#{mvn_cmd} --version", :via => :sudo)
+  }
+
+  task(:test_run_mvn_without_path) {
+    assert_command("mvn --version")
+  }
+
+  task(:test_run_mvn_via_run_locally) {
+    assert_file_not_exists(mvn_bin_local, :via => :run_locally)
+    assert_file_not_exists(File.join(mvn_settings_path_local, "settings.xml"), :via => :run_locally)
+    assert_command_fails("#{mvn_cmd_local} --version", :via => :run_locally)
+  }
+
+  task(:test_mvn_exec) {
+    mvn.exec("--version")
+  }
+
+# task(:test_mvn_exec_locally) {
+#   mvn.exec_locally("--version")
+# }
+
+  task(:test_mvn_artifact) {
+    assert_file_exists(File.join(mvn_project_path, "target", "capistrano-maven-0.0.1-SNAPSHOT.jar"))
+  }
+
+  task(:test_mvn_artifact_locally) {
+    assert_file_not_exists(File.join(mvn_project_path_local, "target", "capistrano-maven-0.0.1-SNAPSHOT.jar"), :via => :run_locally)
+  }
+}
+
+namespace(:test_with_local) {
+  task(:default) {
+    methods.grep(/^test_/).each do |m|
+      send(m)
+    end
+  }
+  before "test_with_local", "test_with_local:setup"
+  after "test_with_local", "test_with_local:teardown"
+
+  task(:setup) {
+    uninstall_mvn!
+    set(:mvn_version, "3.0.5")
+    set(:mvn_skip_tests, true)
+    set(:mvn_setup_remotely, false)
+    set(:mvn_setup_locally, true)
+    set(:mvn_update_remotely, false)
+    set(:mvn_update_locally, true)
+    set(:mvn_template_path, File.join(File.dirname(__FILE__), "templates"))
+    set(:mvn_settings, %w(settings.xml))
+    find_and_execute_task("mvn:setup_default_environment")
+    find_and_execute_task("deploy:setup")
+    find_and_execute_task("deploy")
+  }
+
+  task(:teardown) {
+    uninstall_mvn!
+  }
+
+  task(:test_run_mvn) {
+    assert_file_not_exists(mvn_bin)
+    assert_file_not_exists(File.join(mvn_settings_path, "settings.xml"))
+    assert_command_fails("#{mvn_cmd} --version")
+  }
+
+  task(:test_run_mvn_via_sudo) {
+    assert_command_fails("#{mvn_cmd} --version", :via => :sudo)
+  }
+
+  task(:test_run_mvn_without_path) {
+    assert_command_fails("mvn --version")
+  }
+
+  task(:test_run_mvn_via_run_locally) {
+    assert_file_exists(mvn_bin_local, :via => :run_locally)
+    assert_file_exists(File.join(mvn_settings_path_local, "settings.xml"), :via => :run_locally)
+    assert_command("#{mvn_cmd_local} --version", :via => :run_locally)
+  }
+
+# task(:test_mvn_exec) {
+#   mvn.exec("--version")
+# }
+
+  task(:test_mvn_exec_locally) {
+    mvn.exec_locally("--version")
+  }
+
+  task(:test_mvn_artifact) {
+    assert_file_exists(File.join(mvn_project_path, "target", "capistrano-maven-0.0.1-SNAPSHOT.jar"))
+  }
+
+  task(:test_mvn_artifact_locally) {
     assert_file_exists(File.join(mvn_project_path_local, "target", "capistrano-maven-0.0.1-SNAPSHOT.jar"), :via => :run_locally)
   }
 }
